Revert "fix Settings test and refactor folder structures"
This reverts commit 68fc98b53ad6d2ec55712a2dcb50a04e6c9e1b73.
