+++ apps/settings/elements/phone_lock_passcode.html
@@ -38,7 +38,7 @@
       </div>
     </div>
 
-    <panel data-path="panels/phone_lock_dialog/panel"></panel>
+    <panel data-path="panels/phone_lock/phone_lock_dialog_panel"></panel>
 
   </template>
 </element>
