+++ apps/settings/js/panels/phone_lock/phone_lock_dialog.js
