+++ apps/settings/js/panels/phone_lock/phone_lock_dialog_panel.js
@@ -1,7 +1,7 @@
 define(function(require) {
   'use strict';
   var SettingsPanel = require('modules/settings_panel');
-  var PhoneLockDialog = require('panels/phone_lock_dialog/phone_lock_dialog');
+  var PhoneLockDialog = require('panels/phone_lock/phone_lock_dialog');
 
   return function ctor_phonelockDialog() {
     var phoneLockDialog = PhoneLockDialog();
