+++ apps/settings/test/unit/modules/settings_service_test.js
@@ -13,7 +13,6 @@ mocha.setup({
 
 suite('SettingsService', function() {
   var realL10n;
-  var realSettings;
 
   suiteSetup(function(done) {
     navigator.addIdleObserver = sinon.spy();
@@ -22,8 +21,7 @@ suite('SettingsService', function() {
       'unit/mock_l10n',
       'modules/settings_service',
       'modules/panel_cache',
-      'unit/mock_settings_panel',
-      'settings',
+      'unit/mock_settings_panel'
     ];
 
     // Use map to mock the dependencies.
@@ -39,30 +37,24 @@ suite('SettingsService', function() {
     };
 
     testRequire(modules, map,
-      (function(MockL10n, SettingsService, PanelCache,
-        MockSettingsPanel, Settings) {
-          this.SettingsService = SettingsService;
-          this.PanelCache = PanelCache;
-          // Mock of the SettingsPanel function
-          this.MockSettingsPanel = MockSettingsPanel;
+      (function(MockL10n, SettingsService, PanelCache, MockSettingsPanel) {
+        this.SettingsService = SettingsService;
+        this.PanelCache = PanelCache;
+        // Mock of the SettingsPanel function
+        this.MockSettingsPanel = MockSettingsPanel;
 
-          realSettings = window.Settings;
-          window.Settings = Settings;
+        realL10n = window.navigator.mozL10n;
+        window.navigator.mozL10n = MockL10n;
 
-          realL10n = window.navigator.mozL10n;
-          window.navigator.mozL10n = MockL10n;
-
-          // XXX: As we use 'require' function of requirejs in PanelCache and 
-          //      it conflicts to the original require function, we replace it
-          //      here.
-          this.originalRequire = window.require;
-          window.require = testRequire;
-          done();
+        // XXX: As we use 'require' function of requirejs in PanelCache and it
+        //      conflicts to the original require function, we replace it here.
+        this.originalRequire = window.require;
+        window.require = testRequire;
+        done();
     }).bind(this));
   });
 
   suiteTeardown(function() {
-    window.Settings = realSettings;
     window.navigator.mozL10n = realL10n;
     window.require = this.originalRequire;
     this.originalRequire = null;
